+++ src/main/java/io/enderdev/endermodpacktweaks/core/EMTMixinLoader.java
@@ -41,6 +41,7 @@ public class EMTMixinLoader implements ILateMixinLoader {
             put("mixins.emt.waila.json", () -> Loader.isModLoaded("waila") && EMTConfig.WAILA.enable);
             put("mixins.emt.pickletweaks.json", () -> Loader.isModLoaded("pickletweaks") && EMTConfig.PICKLE_TWEAKS.enable);
             put("mixins.emt.matteroverdrive.json", () -> Loader.isModLoaded("matteroverdrive") && EMTConfig.MATTER_OVERDRIVE.enable);
+            put("mixins.emt.itemstages.json", () -> Loader.isModLoaded("itemstages") && EMTConfig.GAME_STAGES.enable);
         }
     });